requirements.txt
uvicorn>=0.23.2
python-dotenv==0.21.1
requests==2.31.0
pytest==7.4.3
redis[hiredis]==5.0.0
aiohttp==3.9.3
fastapi==0.109.2
pydantic==2.6.1
httpx==0.26.0
Secweb==1.9.1
prometheus-fastapi-instrumentator==6.1.0
rustworkx==0.14.1

